webapi/webapi-webcl-nonw3c-tests/webcl/khronos/conformance/bindingTesting/fetchInfo/cl_program_getBuildInfo_async.html

<!--
 Copyright (C) 2013, 2014 Samsung Electronics Corporation.
 Copyright (C) 2014 Worldline.
 All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLProgram getBuildInfo conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
var callbackFlag1 = false, callbackFlag2 = false;
var webCLProgram1, webCLProgram2, webCLProgram3;
var callback1 = function (event) {
    if (webCLProgram1) {
        callbackFlag1 = true;

        shouldBeNonNull("webCLProgram1.getBuildInfo(webCLDevice1, webcl.PROGRAM_BUILD_STATUS);");
        shouldBe("typeof " + "webCLProgram1.getBuildInfo(webCLDevice1, webcl.PROGRAM_BUILD_OPTIONS);", "'string'");
        shouldBe("typeof " + "webCLProgram1.getBuildInfo(webCLDevice1, webcl.PROGRAM_BUILD_LOG);", "'string'");
    } else
        testFailed("WebCLProgram1 is not a valid object.");
};
var callback2 = function (event) {
    if (webCLProgram2) {
        callbackFlag2 = true;

        var webCLDevice2 = webCLDevices[1];
        shouldThrowExceptionName("webCLProgram1.getBuildInfo(webCLDevice2, webcl.PROGRAM_BUILD_STATUS);", "INVALID_DEVICE");
    } else
        testFailed("WebCLProgram2 is not a valid object.");
};
try {
    debug("WebCLProgram :: getBuildInfo test case.");
    debug("(WebCLDevice device, CLenum name)");

    var wtu = WebCLTestUtils;
    var webCLContext1 = wtu.createContext();
    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);
    var webCLDevice1 = webCLDevices[0];

    var kernelSource = wtu.readKernel("../../../resources/kernels/square.cl");
    var webCLProgram1 = wtu.createProgram(webCLContext1, kernelSource);
    wtu.build(webCLProgram1, webCLDevices, "-cl-single-precision-constant", callback1);

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLProgram1.getBuildInfo(webCLDevice1, null);", "INVALID_VALUE");

    // Testing for huge build log.
    kernelSource = wtu.readKernel("../../../resources/kernels/nbody_simulation_bad.cl");
    webCLProgram3 = wtu.createProgram(webCLContext1, kernelSource);
    try {
        wtu.build(webCLProgram3, webCLDevices);
    } catch(e) {
        shouldBe("typeof " + "webCLProgram3.getBuildInfo(webCLDevice1, webcl.PROGRAM_BUILD_LOG);", "'string'");
    }

    setTimeout(function() {
        if (!callbackFlag1)
            testFailed("webCLProgram1 did not call callback, as expected.");

        // Testing for exception : INVALID_DEVICE.
        if (webCLDevices.length > 1) {
            var webCLDevice1 = webCLDevices[0];
            var webCLContext2 = wtu.createContext(webCLDevice1);

            kernelSource = wtu.readKernel("../../../resources/kernels/square.cl");
            webCLProgram2 = wtu.createProgram(webCLContext2, kernelSource);
            wtu.build(webCLProgram2, [webCLDevice1], '', callback2);
        }  else {
            callbackFlag2 = true;
        }

        setTimeout(function() {
            notifyFinishedToHarness();
            if (!callbackFlag2)
                testFailed("webCLProgram2 did not call callback, as expected.");

            wtu.appendPostJSToHTML(document);
        }, 2000);

    }, 2000);

} catch(e) {
    testFailed(e.description);
    notifyFinishedToHarness();
    wtu.appendPostJSToHTML(document);
}
successfullyParsed = true;
</script>
</body>
</html>
